Put RawPrometheus related files into a separate folder
